change types

+++ BtHomeSensor/Nordic/nrf52840/app/include/interface.h
@@ -7,13 +7,13 @@ extern "C"
 {
 #endif
 
-  int bmp280_i2c_init(uint8_t deviceAddress);
-  int bmp280_i2c_deinit(uint8_t deviceAddress);
-  int bmp280_read_array(uint8_t deviceAddress, uint8_t startRegisterAddress, uint8_t *data, uint8_t dataLength);
-  int bmp280_write_array(uint8_t deviceAddress, uint8_t startRegisterAddress, uint8_t *data, uint8_t dataLength);
-  int delay_function(uint32_t delayMS);
-  int power_function(float x, float y, float *result);
-  int getMode();
+  int8_t bmp280_i2c_init(uint8_t deviceAddress);
+  int8_t bmp280_i2c_deinit(uint8_t deviceAddress);
+  int8_t bmp280_read_array(uint8_t deviceAddress, uint8_t startRegisterAddress, uint8_t *data, uint8_t dataLength);
+  int8_t bmp280_write_array(uint8_t deviceAddress, uint8_t startRegisterAddress, uint8_t *data, uint8_t dataLength);
+  int8_t delay_function(uint32_t delayMS);
+  int8_t power_function(float x, float y, float *result);
+  int8_t getMode();
 
 #ifdef __cplusplus
 }
